Fixed interest rate - now a float
Previously we were checking interest rate as a int instead of float.
Added float interest rate scenarios to the test.

+++ src/main/java/com/mortgagecalc/pages/CalculatorPage.java
@@ -44,7 +44,7 @@ public class CalculatorPage extends BasePage {
 		selectByValue(termInYears, termInYearsSelect);
 	}
 
-	public void setInterestRate(int interestRate) {
+	public void setInterestRate(float interestRate) {
 		type(String.valueOf(interestRate), interestRateFieldYear);
 	}
